change default random generator to std::mt19937.

+++ src/tthreadapplicationserver.cpp
@@ -74,7 +74,7 @@ void TThreadApplicationServer::incomingConnection(
 #endif
 {
     for (;;) {
-        tSystemDebug("incomingConnection  sd:%lld  thread count:%d  max:%d", socketDescriptor, TActionThread::threadCount(), maxThreads);
+        tSystemDebug("incomingConnection  sd:%lld  thread count:%d  max:%d", (qint64)socketDescriptor, TActionThread::threadCount(), maxThreads);
         if (TActionThread::threadCount() < maxThreads) {
             TActionThread *thread = new TActionThread(socketDescriptor);
             connect(thread, SIGNAL(finished()), thread, SLOT(deleteLater()));
